\name{gloptim-package}
\alias{gloptim-package}
\alias{gloptim}
\docType{package}
\title{
  Global Optimization
}
\description{
  The R package `gloptim' will provide methods and algorithms for 
  global optimization and function minimization of nonlinear functions. 
  Derivative-free, stochastic, and noise-tolerant approaches will be 
  included.
}
\details{
\tabular{ll}{
Package: \tab gloptim\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2011-07-27\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\author{
  Hans W Borchers

  Maintainer: HwB  <hwborchers@googlemail.com>
}
\references{
  Arnold Neumaier (2010). Global Optimization.
  \url{http://www.mat.univie.ac.at/~neum/glopt.html}
}
\keyword{ package }
\seealso{
  The R-Forge project ``Optimization and solving packages''.
  \url{https://r-forge.r-project.org/projects/optimizer/}
}
\examples{
# See the functions in this package.
}
